requirements.txt
pandas>=2.0
yfinance>=0.2.40
python-dotenv>=1.0
playwright>=1.40
requests>=2.31
